Add CameraManager to Gameplay PlayerController, update WBPs
Headbob CameraModifier was missing and it was in control of the footstep cycle from the footstep component. With the Cameramanager added back to the gameplay PC, the checking is now back.
Updated GameState and GameMode to handle communication for Missions, as the GameState now handles tracking the time and GameMode will handle telling the GameState if the player has won or failed.